backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-core==2.20.1
scikit-learn==1.6.1
joblib==1.4.2
python-multipart==0.0.9
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<1.13.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
alembic==1.13.2
email-validator==2.2.0 